requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
huggingface_hub>=0.24.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
ddgs>=7.0.0
